requirements.txt
appdirs==1.4.4
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
docker-pycreds==0.4.0
filelock @ file:///croot/filelock_1672387128942/work
fonttools==4.43.1
fsspec==2023.10.0
ftfy==6.1.1
gitdb==4.0.11
GitPython==3.1.40
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
huggingface-hub==0.18.0
idna @ file:///croot/idna_1666125576474/work
importlib-resources==6.1.0
Jinja2 @ file:///croot/jinja2_1666908132255/work
kiwisolver==1.4.5
lightning-utilities==0.9.0
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.3
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpire==2.8.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
Pillow @ file:///croot/pillow_1696580024257/work
protobuf==4.24.4
psutil==5.9.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.2
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex==2023.10.3
requests @ file:///croot/requests_1690400202158/work
rich==13.6.0
safetensors==0.4.0
scipy==1.10.1
sentry-sdk==1.33.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
spicy==0.16.0
sympy @ file:///croot/sympy_1668202399572/work
timm==0.9.8
torch==2.1.0
torchaudio==2.1.0
torchmetrics==0.8.2
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1698257533958/work
wandb==0.15.12
wcwidth==0.2.9
zipp==3.17.0
